apply an hack
it seems like --follow fails because container is not yet created, e.g container has no logs. so what I'll do is just listen for the logs in the background and wait for the job to complete on the foreground.
there is still an open issue on kubernetes community on github sense 2016 https://github.com/kubernetes/kubernetes/issues/28746